// src/network/api/model/alarmLocationTimeRequest.ts
/**
 * Generated by orval v6.18.1 🍺
 * Do not edit manually.
 * Springdoc
 * Springdoc을 사용한 Swagger UI
 * OpenAPI spec version: 1.0.0
 */

/**
 * 알람지역시간목록
 */
export interface AlarmLocationTimeRequest {
  /** 지역시간ID */
  alarmLocationTimeId?: number;
  /** 지역시간 */
  locationTime?: string;
}