// src/network/api/model/alarmUpdateRequest.ts
/**
 * Generated by orval v6.18.1 🍺
 * Do not edit manually.
 * Springdoc
 * Springdoc을 사용한 Swagger UI
 * OpenAPI spec version: 1.0.0
 */
import type { AlarmWeekRequest } from "./alarmWeekRequest";
import type { AlarmLocationRequest } from "./alarmLocationRequest";

export interface AlarmUpdateRequest {
  /** 유저ID */
  userId?: number;
  /** 알람ID */
  alarmId?: number;
  /** 시간(오후 1시30분이면 0130) */
  alarmTime?: string;
  /** AM 또는 PM */
  timeOfDay?: string;
  /** 다시울림시간(1분부터59분까지) */
  reminder?: string;
  /** 진동여부(0:없음, 1:진동) */
  vibration?: number;
  /** 요일목록 */
  weekList?: AlarmWeekRequest[];
  /** 알람소리코드 */
  alarmSoundId?: number;
  /** 볼륨(1부터 10까지) */
  volume?: number;
  /** 알람지역목록 */
  alarmLocationList?: AlarmLocationRequest[];
}