// src/network/api/model/alarmWeekRequest.ts
/**
 * Generated by orval v6.18.1 🍺
 * Do not edit manually.
 * Springdoc
 * Springdoc을 사용한 Swagger UI
 * OpenAPI spec version: 1.0.0
 */

/**
 * 요일목록
 */
export interface AlarmWeekRequest {
  /** 알람요일ID */
  alarmWeekId?: number;
  /** 요일코드(1부터 월요일) */
  weekId?: number;
}